Refactor: Consolidate Loader component and update modal usage
- Replaced legacy Loader imports with a unified Loader component across various files.
- Updated modal implementations in AuthorizedNav, BalanceBar, and removed deprecated modal components.
- Improved styling and structure in Loader and modal components for better consistency.
