+++ src/components/Loader/Loader.styled.ts
