+++ src/components/Loader/Loader.tsx
@@ -1,7 +1,6 @@
 // import { ThreeCircles } from "react-loader-spinner";
 import { LoaderWrapper } from "./Loader.styled";
 
-// Should this reside in a "service" folder? It's definitely a component :)
 const Loader = () => {
   return (
     <LoaderWrapper>
