+++ src/components/Modals/index.ts
@@ -2,7 +2,3 @@
 export { Modal } from "./Modal";
 export type { ModalVariant } from "./Modal";
 
-// Legacy Components (now using unified Modal internally)
-export { HeaderModalWindow } from "./HeaderModalWindow/HeaderModalWindow";
-export { default as BalanceModal } from "./BalanceModal/BalanceModal";
-export { default as EntryModal } from "./EntryModal/EntryModal";
