+++ src/pages/SharedLayout/SharedLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import { Suspense } from "react";
-import Loader from "../../service/Loader/Loader";
+import Loader from "../../components/Loader/Loader";
 import Header from "../../components/Header/Header";
 import Footer from "../../components/Footer/Footer";
 import { useMatchMedia } from "../../utils/hooks/useMatchMedia";
